llama_script/requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
filelock==3.16.1
fsspec==2024.10.0
huggingface-hub==0.26.2
idna==3.10
lxml==5.3.0
numpy==2.1.3
packaging==24.2
PyMuPDF==1.24.13
python-docx==1.1.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.4.5
tokenizers==0.20.3
tqdm==4.67.0
transformers==4.46.2
typing_extensions==4.12.2
urllib3==2.2.3
